update sale status on negotiation creation
