+++ libs/contract/src/lib/negotiation/+state/negotiation.service.ts
@@ -3,9 +3,8 @@ import { EntityState, ActiveState, StoreConfig, EntityStore } from "@datorama/ak
 import { CollectionConfig, CollectionService } from "akita-ng-fire";
 import { Negotiation } from "./negotiation.firestore";
 import type firebase from 'firebase';
-import { createDocumentMeta, formatDocumentMetaFromFirestore } from "@blockframes/utils/models-meta";
+import {  formatDocumentMetaFromFirestore } from "@blockframes/utils/models-meta";
 import { OrganizationQuery } from "@blockframes/organization/+state";
-import { centralOrgId } from "@env";
 
 export interface NegotiationState extends EntityState<Negotiation, string>, ActiveState<string> { }
 
@@ -18,27 +17,6 @@ export class NegotiationService extends CollectionService<NegotiationState> {
     private orgQuery: OrganizationQuery,
   ) {
     super(store)
-  }
-
-  create(contractId: string, contract: Partial<Negotiation>) {
-    const activeOrgId = this.orgQuery.getActiveId();
-    return this.add({
-      _meta: createDocumentMeta({ createdAt: new Date(), }),
-      status: 'pending',
-      createdByOrg: activeOrgId,
-      sellerId: centralOrgId.catalog,
-      stakeholders: contract.stakeholders,
-      buyerId: contract.orgId,
-      price: contract.price,
-      titleId: contract.titleId,
-      terms: contract.terms,
-      holdbacks: contract.holdbacks,
-      parentTermId: contract.parentTermId,
-      specificity: contract.specificity,
-      orgId: contract.orgId,
-    }, {
-      params: { contractId }
-    })
   }
 
   formatFromFirestore(_negotiation: Negotiation<firebase.firestore.Timestamp>): Negotiation<Date> {
